requirements_new.txt
asgiref             3.7.2  
backports.zoneinfo  0.2.1  
Django              4.2.4  
djangorestframework 3.14.0 
djongo              1.3.6  
dnspython           2.4.2  
pip                 20.0.2 
pkg-resources       0.0.0  
psycopg2-binary     2.9.7  
pymongo             4.5.0  
pytz                2023.3 
setuptools          44.0.0 
sqlparse            0.4.4  
typing-extensions   4.7.1  
tzdata              2023.3 
wheel               0.34.2 
